+++ df-qol/src/df-qol.ts
@@ -82,6 +82,30 @@ Hooks.once('init', function () {
 	});
 	game.settings.settings.get(`${SETTINGS.MOD_NAME}.better-toggle`).onChange(SETTINGS.get('better-toggle'));
 
+	SETTINGS.register('day-night-progress', {
+		scope: 'world',
+		config: true,
+		type: Boolean,
+		default: true,
+		name: 'DRAGON_FLAGON_QOL.DayNightProgress_SettingName',
+		hint: 'DRAGON_FLAGON_QOL.DayNightProgress_SettingHint',
+		onChange: newValue => {
+			libWrapper.unregister(SETTINGS.MOD_NAME, 'LightingLayer.prototype.animateDarkness');
+		}
+	});
+	if (SETTINGS.get('day-night-progress')) {
+		libWrapper.register(SETTINGS.MOD_NAME, 'LightingLayer.prototype.animateDarkness', DF_DAY_NIGHT_PROGRESS, 'OVERRIDE');
+	}
+	SETTINGS.register('day-night-duration', {
+		scope: 'world',
+		config: true,
+		type: Number,
+		default: 10,
+		range: { min: 1, max: 60, step: 1 },
+		name: 'DRAGON_FLAGON_QOL.DayNightDuration_SettingName',
+		hint: 'DRAGON_FLAGON_QOL.DayNightDuration_SettingHint'
+	});
+
 	// If we have D&D5e running
 	if (!!game.dnd5e) {
 		SETTINGS.register('vehicle-unit', {
@@ -98,6 +122,12 @@ Hooks.once('init', function () {
 });
 
 Hooks.once('ready', () => {
+	if (!game.modules.get('lib-wrapper')?.active) {
+		console.error('Missing libWrapper module dependency');
+		if (game.user.isGM)
+			ui.notifications.error(game.i18n.localize('DF-QOL.errorLibWrapperMissing'));
+		return;
+	}
 	// If we have D&D5e running
 	if (!!game.dnd5e)
 		DF_VEHICLE_UNITS();
@@ -257,4 +287,46 @@ function DF_VEHICLE_UNIT_CONFIG(app: EntitySheetConfig, html: JQuery<HTMLElement
 		await (<Actor>this.object).setFlag(SETTINGS.MOD_NAME, 'unit', unit);
 		return core(event, formData);
 	}
+}
+
+async function DF_DAY_NIGHT_PROGRESS(this: LightingLayer, target = 1.0, { duration = 10000 } = {}) {
+	const animationName = "lighting.animateDarkness";
+	CanvasAnimation.terminateAnimation(animationName);
+	if (target === this.darknessLevel) return false;
+	if (duration <= 0) return this.refresh(target);
+
+	if (duration === 10000)
+		duration = SETTINGS.get('day-night-duration') as number * 1000;
+	// Prepare the animation data object
+	this._animating = true;
+	const animationData = [{
+		parent: { darkness: this.darknessLevel },
+		attribute: "darkness",
+		to: Math.clamped(target, 0, 1)
+	}];
+
+	// Trigger the animation function
+	let elapsed = 0.0;
+	let last = new Date().getTime();
+	return CanvasAnimation.animateLinear(animationData, {
+		name: animationName,
+		duration: duration,
+		context: undefined,
+		ontick: (dt, attributes) => {
+			this.refresh(attributes[0].parent.darkness);
+			// do not display the transition progress to the PCs, only to GMs
+			if (!game.user.isGM) return;
+			// show progress here
+			const now = new Date().getTime();
+			elapsed += now - last;
+			last = now;
+			const loader = document.getElementById("loading");
+			const pct = Math.ceil((elapsed / duration) * 100);
+			loader.querySelector("#context").textContent = 'Day/Night Transitioning...';
+			(loader.querySelector("#loading-bar") as HTMLElement).style.width = `${pct}%`;
+			loader.querySelector("#progress").textContent = `${Math.round(elapsed / 1000)}/${Math.ceil(duration / 1000)} sec`;
+			loader.style.display = "block";
+			if ((duration - elapsed < 500) && !loader.hidden) $(loader).fadeOut(2000);
+		}
+	});
 }